requirements.txt
django>=5.0.1
Pillow
djangorestframework
pyyaml
requests
django-cors-headers
drf-yasg
setuptools